break down the list to smaller components
